+++ Local/src/main/java/ir/omidtaheri/local/datasource/MovieLocalDataSourceImp.kt
@@ -17,7 +17,7 @@ class MovieLocalDataSourceImp @Inject constructor(
         return movieDao.favoriteMovie(movieEntityDataLocalMapper.mapFromDataEntity(movie))
     }
 
-    override suspend fun unFavoriteMovie(movie: FavoritedMovieDataEntity): Int {
+    override suspend fun unFavoriteMovie(movie: FavoritedMovieDataEntity): Long {
         return movieDao.unFavoriteMovie(movieEntityDataLocalMapper.mapFromDataEntity(movie))
     }
